estimates_model/requirements.txt
appnope==0.1.2
backcall==0.2.0
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cycler==0.10.0
decorator==5.0.7
envyaml==1.8.210417
google-api-core==1.26.3
google-auth==1.30.0
google-cloud-bigquery==2.16.0
google-cloud-core==1.6.0
google-cloud-firestore==2.1.1
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.37.1
idna==2.10
ipython-genutils==0.2.0
jedi==0.18.0
kiwisolver==1.3.1
matplotlib-inline==0.1.2
numpy==1.20.3
packaging==20.9
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
proto-plus==1.18.1
protobuf==3.16.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.9.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
scipy==1.6.3
six==1.16.0
torch==1.8.1
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.4
utm==0.7.0
wcwidth==0.2.5
